day2 v0.0.1 Day2

Helper functions for day2 of Advent of code

Types

dimensions :: {integer, integer, integer}

Tuple representing length, width, and height

Functions

get_paper_amount(text)

Specs

get_paper_amount(String.t) :: integer

Takes a string representation of a box and returns the amount of necessary wrapping paper

#Examples

iex> Day2.get_paper_amount("2x3x4")
  58

  iex> Day2.get_paper_amount("1x1x10")
  43

get_ribbon_amount(text)

Specs

get_ribbon_amount(String.t) :: integer

Takes a string representation of a box and returns the amount of necessary ribbon

#Examples

iex> Day2.get_ribbon_amount("2x3x4")
  34

  iex> Day2.get_ribbon_amount("1x1x10")
  14