day5 v0.0.1 Day5

Helper functions for day5 of Advent of code

Functions

contains_dups?(text)

Specs

contains_dups?(String.t) :: boolean

Checks if text contains adjacent duplicate letter

#Examples

iex> Day5.contains_dups?("ugknbfddgicrmopn")
true

iex> Day5.contains_dups?("aaa")
true

iex> Day5.contains_dups?("jchzalrnumimnmhp")
false

contains_forbidden?(text)

Specs

contains_forbidden?(String.t) :: boolean

Checks if text contains forbidden sequences

#Examples

iex> Day5.contains_forbidden?("ugknbfddgicrmopn")
false

iex> Day5.contains_forbidden?("aaa")
false

iex> Day5.contains_forbidden?("haegwjzuvuyypxyu")
true

contains_spaced_dup?(text)

Specs

contains_spaced_dup?(String.t) :: boolean

Checks if a character repeats with a single character between the repetition

#Examples #Examples

iex> Day5.contains_spaced_dup?("qjhvhtzxzqqjkmpb")
true

iex> Day5.contains_spaced_dup?("xxyxx")
true

iex> Day5.contains_spaced_dup?("uurcxstgmygtbstg")
false

iex> Day5.contains_spaced_dup?("ieodomkazucvgmuy")
true

has_duplicate?(members)

Specs

has_duplicate?(Enumerable.t) :: boolean

Checks if Enumerable contains any duplicate members

#Examples

iex> Day5.has_duplicate?(["a", 1, 3, "a"])
true

iex> Day5.has_duplicate?([1, 2, 3, 4])
false

is_nice?(text)

Specs

is_nice?(String.t) :: boolean

Determines if a string is nice

#Examples

iex> Day5.is_nice?("ugknbfddgicrmopn")
true

iex> Day5.is_nice?("aaa")
true

iex> Day5.is_nice?("jchzalrnumimnmhp")
false

iex> Day5.is_nice?("haegwjzuvuyypxyu")
false

iex> Day5.is_nice?("dvszwmarrgswjxmb")
false

repeated_tuple?(text)

Specs

repeated_tuple?(String.t) :: boolean

Checks if text contains any repeated non-overlapping tuples

#Examples

iex> Day5.repeated_tuple?("qjhvhtzxzqqjkmpb")
true

iex> Day5.repeated_tuple?("xxyxx")
true

iex> Day5.repeated_tuple?("uurcxstgmygtbstg")
true

iex> Day5.repeated_tuple?("ieodomkazucvgmuy")
false

super_nice?(text)

Specs

super_nice?(String.t) :: boolean

Determines if a string is nice according to the year 2 rules

#Examples

iex> Day5.super_nice?("qjhvhtzxzqqjkmpb")
true

iex> Day5.super_nice?("xxyxx")
true

iex> Day5.super_nice?("uurcxstgmygtbstg")
false

iex> Day5.super_nice?("ieodomkazucvgmuy")
false

three_vowels?(text)

Specs

three_vowels?(String.t) :: boolean

Checks if text has three vowels

#Examples

iex> Day5.three_vowels?("ugknbfddgicrmopn")
true

iex> Day5.three_vowels?("aaa")
true

iex> Day5.three_vowels?("dvszwmarrgswjxmb")
false