day_1 v0.0.1 UnbalancedParen

Module for dealing with strings with an unbalanced set of parentheses

Functions

get_balance(text)

Specs

get_balance(String.t) ::
  {:balanced, 0} |
  {:negative, integer} |
  {:positive, integer}

Get the balance for the string of parentheses, failing fast if degree ever goes negative. If the result is a negative balance, the failure location will be returned. Otherwise the degree will be returned.

iex> UnbalancedParen.get_balance(")")
{:negative, 1}

iex> UnbalancedParen.get_balance("()())")
{:negative, 5}

iex> UnbalancedParen.get_balance("()(")
{:positive, 1}

iex> UnbalancedParen.get_balance("(()())")
{:balanced, 0}

get_degree(text)

Specs

get_degree(String.t) :: integer

Get the degree of unbalance for the string of parentheses

#Examples

iex> UnbalancedParen.get_degree("(())")
0

iex> UnbalancedParen.get_degree("()()")
0

iex> UnbalancedParen.get_degree("(((")
3

iex> UnbalancedParen.get_degree("(()(()(")
3

iex> UnbalancedParen.get_degree("))(((((")
3

iex> UnbalancedParen.get_degree("())")
-1

iex> UnbalancedParen.get_degree("))(")
-1

iex> UnbalancedParen.get_degree(")))")
-3

iex> UnbalancedParen.get_degree(")())())")
-3